Move the events from models packages
